pyproject.toml
[tool.poetry]
name = "fastapi-tutorial"
version = "0.1.0"
description = ""
authors = ["Kawaguchi Makoto"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
pydantic = "^1.8.2"
python-ulid = "^1.0.3"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.29"}
aiomysql = "^0.0.22"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
pyproject-flake8 = "^0.0.1-alpha.2"
mypy = "^0.930"
types-PyMySQL = "^1.0.7"
requests = "^2.26.0"
httpx = "^0.21.1"
pytest-asyncio = "^0.16.0"
aiosqlite = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Following configurations are added manually

[tool.black]
line-length = 119
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
ignore="E203,W503,W504"
max-line-length=119
max-complexity=10

[tool.mypy]
# common
python_version = 3.9
show_column_numbers  = true
show_error_context = true
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
# warning
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true

[tool.isort]
profile = "black"
line_length = 119
